+++ vignettes/statgenSTA.R
@@ -80,7 +80,8 @@ plot(wheatTD, plotType = "cor", traits = "GY")
 
 ## ----scatPlot, fig.dim = c(7, 7)----------------------------------------------
 ## Create a scatter plot matrix for grain yield.
-plot(wheatTD, plotType = "scatter", traits = "GY")
+## Add correlations between trials in top left of scatter plots.
+plot(wheatTD, plotType = "scatter", traits = "GY", addCorr = "tl")
 
 ## ----fitSp, message=FALSE-----------------------------------------------------
 ## Fit a single trial model using a model based on a residual row column design.