+++ NEWS.md
@@ -1,3 +1,8 @@
+# statgenSTA 1.0.3
+
+* It is now possible to create a scatter plot matrix for objects of class TD.
+* statgenSTA is no longer dependent on reshape2.
+
 # statgenSTA 1.0.2
 
 * Initial CRAN version